swfs-project-backend/requirements.txt
Flask
dnspython
flask-cors
pymongo
twilio
python-dotenv
waitress
pytz

